requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
django-crispy-forms==2.0
django-extensions==3.2.1
Django==4.1.7
factory-boy==3.2.1
Faker==17.3.0
pip==21.3.1
psycopg2==2.9.5
python-dateutil==2.8.2
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
